Refactor osm-module (still broken)
